pdf-to-suggestion/requirements.txt
streamlit
google-cloud-aiplatform
python-dotenv
